requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
black==21.7b0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
gitdb==4.0.7
GitPython==3.1.20
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
regex==2021.7.6
smmap==4.0.0
tomli==1.1.0
typing-extensions==3.10.0.0
yarl==1.6.3
